Consolidate registration queue
